fix circular dependencies (1)
