+++ apps/fishing-map/features/reports/activity/ReportActivityBeforeAfterGraph.tsx
@@ -16,7 +16,7 @@ import { Interval } from '@globalfishingwatch/layer-composer'
 import { selectReportTimeComparison } from 'features/app/app.selectors'
 import { ReportActivityTimeComparison } from 'types'
 import i18n from 'features/i18n/i18n'
-import { COLOR_PRIMARY_BLUE } from 'features/app/App'
+import { COLOR_PRIMARY_BLUE } from 'features/app/app.config'
 import { getUTCDateTime } from 'utils/dates'
 import { formatDate, tickFormatter } from 'features/reports/reports.utils'
 import { formatI18nNumber } from 'features/i18n/i18nNumber'
