+++ apps/fishing-map/features/reports/activity/ReportActivityPeriodComparisonGraph.tsx
@@ -18,7 +18,7 @@ import {
   selectReportTimeComparison,
 } from 'features/app/app.selectors'
 import i18n, { t } from 'features/i18n/i18n'
-import { COLOR_GRADIENT, COLOR_PRIMARY_BLUE } from 'features/app/App'
+import { COLOR_GRADIENT, COLOR_PRIMARY_BLUE } from 'features/app/app.config'
 import { getUTCDateTime } from 'utils/dates'
 import { formatDate, formatTooltipValue, tickFormatter } from 'features/reports/reports.utils'
 import { EMPTY_FIELD_PLACEHOLDER } from 'utils/info'
