+++ apps/fishing-map/features/reports/environment/ReportEnvironment.tsx
@@ -5,7 +5,7 @@ import { DatasetTypes } from '@globalfishingwatch/api-types'
 import { selectActiveReportDataviews, selectTimeRange } from 'features/app/app.selectors'
 import { useFilteredTimeSeries } from 'features/reports/reports-timeseries.hooks'
 import ReportActivityPlaceholder from 'features/reports/placeholders/ReportActivityPlaceholder'
-import { getDatasetNameTranslated } from 'features/i18n/utils'
+import { getDatasetNameTranslated } from 'features/i18n/utils.datasets'
 import { formatI18nDate } from 'features/i18n/i18nDate'
 import ReportActivityEvolution from '../activity/ReportActivityEvolution'
 import styles from './ReportEnvironment.module.css'
