+++ apps/fishing-map/features/workspace/common/InfoModalContent.tsx
@@ -2,7 +2,7 @@ import { Fragment } from 'react'
 import parse from 'html-react-parser'
 import { useSelector } from 'react-redux'
 import { Dataset } from '@globalfishingwatch/api-types'
-import { getDatasetDescriptionTranslated } from 'features/i18n/utils'
+import { getDatasetDescriptionTranslated } from 'features/i18n/utils.datasets'
 import { isGFWUser } from 'features/user/user.slice'
 import GFWOnly from 'features/user/GFWOnly'
 import styles from './InfoModal.module.css'
