+++ apps/fishing-map/features/workspace/environmental/UserTrackLayerPanel.tsx
@@ -17,7 +17,7 @@ import { useAutoRefreshImportingDataset } from 'features/datasets/datasets.hook'
 import { selectResourceByUrl } from 'features/resources/resources.slice'
 import { selectActiveTrackDataviews } from 'features/dataviews/dataviews.slice'
 import { useLayerPanelDataviewSort } from 'features/workspace/shared/layer-panel-sort.hook'
-import { COLOR_SECONDARY_BLUE } from 'features/app/App'
+import { COLOR_SECONDARY_BLUE } from 'features/app/app.config'
 import DatasetNotFound from '../shared/DatasetNotFound'
 import Color from '../common/Color'
 import LayerSwitch from '../common/LayerSwitch'
